A-Assertions: Use Assertions
There is no assertions used in the current project to maintain
assumptions as we progress along with the project.

Assumptions in code are not maintained, and could be violated as we
progress.

Assertions are added into the project, to ensure class, control-flow and
internal invariants. Example: Adding null checks to class attributes,
for Deadline.java.

These are the conventions and guidelines Oracle has provided when it
comes to using Assertions statement.